requirements.txt
fastapi
uvicorn
google.genai
PyPDF2
python-dotenv
python-multipart
google.generativeai